Affine cipher in C
the program asks if you want to encrypt the message in the m.txt file or decrypt the ciphertext in the c.txt file. Then, asks an alpha and a beta to cipher or decrypt the text and rewrite in the respective file (m to c... c to m). The program uses modulo 26 because of the english alphabet... but the web version uses any modulo